Move remove_indentation to nicegui.helpers
The helper is no longer markdown-specific now that ui.mermaid uses it
too, so promoting it to nicegui.helpers keeps imports clean and avoids
cross-element dependencies.
