+++ nicegui/helpers/__init__.py
@@ -10,7 +10,7 @@ from .functions import (
     should_await,
 )
 from .network import format_url, is_port_open, schedule_browser
-from .strings import event_type_to_camel_case, kebab_to_camel_case
+from .strings import event_type_to_camel_case, kebab_to_camel_case, remove_indentation
 from .warnings import warn_once
 
 __all__ = [
@@ -26,6 +26,7 @@ __all__ = [
     'is_user_simulation',
     'kebab_to_camel_case',
     'normalize_lifecycle_handler',
+    'remove_indentation',
     'require_top_level_layout',
     'schedule_browser',
     'should_await',
